% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ve_deactivate.R
\name{ve_deactivate}
\alias{ve_deactivate}
\title{Deactivate current virtual environment}
\usage{
ve_deactivate()
}
\description{
Upon deactivation, R environment is set to default package environemnt.
All attached packages are unloaded
}
