% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ve_delete.R
\name{ve_delete}
\alias{ve_delete}
\title{Delete an R environment}
\usage{
ve_delete(names)
}
\arguments{
\item{names}{character vector name of the environment to be deleted.}
}
\description{
Delete an R environment
}
