% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ve_export.R
\name{ve_export}
\alias{ve_export}
\title{Export R package environment}
\usage{
ve_export(name)
}
\arguments{
\item{name}{character name of the environment to be generated}
}
\description{
Generate a .yaml file with full package listings and corresponding versions.
The file can then be used to create new virtualenv settings.
}
