% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ve_new.R
\name{ve_new}
\alias{ve_new}
\title{Initialize a virtual environment}
\usage{
ve_new(name, dependencies)
}
\arguments{
\item{name}{character name of the new environment}

\item{dependancies}{path the yaml config file}
}
\description{
Initialize a virtual environment
}
