% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ve_remove.R
\name{ve_remove}
\alias{ve_remove}
\title{Delete an R environment}
\usage{
ve_remove(names)
}
\arguments{
\item{names}{character vector name of the environment to be deleted.}
}
\description{
Delete an R environment
}
