% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ve_new.R
\name{ve_new}
\alias{ve_new}
\title{Initialize a virtual environment}
\usage{
ve_new(name = NULL, config_path = NULL)
}
\arguments{
\item{name}{character string specifying the name of a new environment}

\item{config_path}{path the yaml config file}
}
\description{
Initialize a virtual environment
}
